refactor: reorganize Supabase client imports and update file structure
