+++ prisma.config.ts
@@ -1,3 +1,4 @@
+import 'dotenv/config';
 import path from "node:path";
 import { defineConfig } from "prisma/config";
