+++ src/app/api/auth/password-reset/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { createAdminClient } from "~/supabase/admin";
+import { createAdminClient } from "@/utils/supabase/admin";
 
 import { resetSchema } from "@/schemas/resetSchema";
