--- supabase/component.ts
@@ -1,10 +1,0 @@
-import { createBrowserClient } from "@supabase/ssr";
-
-export function createClient() {
-  const supabase = createBrowserClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY!
-  );
-
-  return supabase;
-}
