--- supabase/static-props.ts
@@ -1,10 +1,0 @@
-import { createClient as createClientPrimitive } from "@supabase/supabase-js";
-
-export function createClient() {
-  const supabase = createClientPrimitive(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY!
-  );
-
-  return supabase;
-}
